refactor(money): last Double money fields to BigDecimal — slice 5, migration closed

Bundle.discountPercentage, PromoCampaign.discountPercent, SubscriptionPlan.amount
and ReturnRequest.refundAmount move to BigDecimal / NUMERIC(12,2) via V30. Closes
the money migration: no column or field holds money as a binary float, and the
asLegacyDouble / (long)(x*100) boundaries are gone. Full suite 447 green.

+++ ecom-backend/src/main/java/com/ecommerce/project/payload/ReturnRequestDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -23,5 +24,5 @@ public class ReturnRequestDTO {
     private String carrierName;
     private String trackingStatus;
     private LocalDateTime lastTrackedAt;
-    private Double refundAmount;
+    private BigDecimal refundAmount;
 }
